Add Seeds

+++ Models/FaktureProducts.cs
@@ -5,7 +5,7 @@ namespace Models
         public string Name { get; set; }
         public int Count { get; set; }
         public double Price { get; set; }
-        public float Tax { get; set; }
+        public double Tax { get; set; }
 
     }
 }
